Edycja starych list zadań z semestru 1

+++ Semestr 1/Programowanie-1/Lista_1/Zadanie_3.py
@@ -1,5 +1,5 @@
 import math
 
-r = int(5)
+r = float(5)
 objetosc = 4/3 * math.pi * r**3
 print(objetosc)
